requirements.txt
requests==2.25.1
pytest==6.2.2
flask_cors==3.0.10
Flask==1.1.2
Werkzeug==1.0.1
Pillow==8.2.0
PyJWT==2.0.1
